refactor: festival-button 관련 쿼리 제거 및 기존 쿼리에 추가
