+++ apps/client/src/pages/time-table/hooks/use-festival-data.ts
@@ -1,11 +1,10 @@
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query';
 
-import { FESTIVAL_BUTTON_QUERY_OPTIONS } from '@shared/apis/time-table/festival-button-queries';
 import { FESTIVAL_TIMETABLE_QUERY_OPTIONS } from '@shared/apis/time-table/festival-timetable-queries';
 
 export const useFestivalButtonData = () => {
   const { data } = useSuspenseQuery(
-    FESTIVAL_BUTTON_QUERY_OPTIONS.FESTIVAL_BUTTON(),
+    FESTIVAL_TIMETABLE_QUERY_OPTIONS.AVAILABLE_FESTIVALS(),
   );
   return data;
 };
