+++ apps/client/src/pages/time-table/hooks/use-timetable-festival-mutation.ts
@@ -1,11 +1,10 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { addFestivalTimeTable } from '@shared/apis/time-table/festival-button-queries';
-import { deleteFestivalTimetables } from '@shared/apis/time-table/festival-timetable-queries';
 import {
-  FESTIVAL_BUTTON_QUERY_KEY,
-  FESTIVAL_TIMETABLE_QUERY_KEY,
-} from '@shared/constants/query-key';
+  deleteFestivalTimetables,
+  postAddFestivalTimeTable,
+} from '@shared/apis/time-table/festival-timetable-queries';
+import { FESTIVAL_TIMETABLE_QUERY_KEY } from '@shared/constants/query-key';
 
 export const useDeleteTimeTableFestival = () => {
   const queryClient = useQueryClient();
@@ -14,10 +13,7 @@ export const useDeleteTimeTableFestival = () => {
     mutationFn: (festivalId: number) => deleteFestivalTimetables(festivalId),
     onSuccess: () => {
       queryClient.invalidateQueries({
-        queryKey: [
-          ...FESTIVAL_BUTTON_QUERY_KEY.ALL,
-          ...FESTIVAL_TIMETABLE_QUERY_KEY.ALL,
-        ],
+        queryKey: [...FESTIVAL_TIMETABLE_QUERY_KEY.ALL],
       });
     },
   });
@@ -28,10 +24,10 @@ export const useAddTimeTableFestival = (onSuccessCallback?: () => void) => {
 
   return useMutation({
     mutationFn: (selectedFestivals: number[]) =>
-      addFestivalTimeTable(selectedFestivals),
+      postAddFestivalTimeTable(selectedFestivals),
     onSuccess: async () => {
       await queryClient.invalidateQueries({
-        queryKey: [...FESTIVAL_BUTTON_QUERY_KEY.ALL],
+        queryKey: [...FESTIVAL_TIMETABLE_QUERY_KEY.ALL],
       });
       if (onSuccessCallback) {
         onSuccessCallback();
